// src/app/shared/types/api-types/bidt-valuation-group-dto.interface.ts
export interface BidtValuationGroupDTO {
  id?: number;
  valuationGroupCode?: string;
  valuationGroupName?: string;
  rangeValueMin?: string;
  rangeValueMax?: string;
  unitCode?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  rangeValueMinCsv?: string;
  rangeValueMaxCsv?: string;
}